support storage api for query
